Move scarb notifications from scarb.rs to ext.rs
